+++ my_gometry/segment.h
@@ -9,8 +9,8 @@ namespace geometry {
    private:
     Point a1_;
     Point a2_;
-    int64_t x_ = 0;
-    int64_t y_ = 0;
+    float x_ = 0;
+    float y_ = 0;
 
    public:
     Segment();
@@ -19,7 +19,7 @@ namespace geometry {
     Point Geta2() const;
     Segment& Move(const Vector&) override;
     bool ContainsPoint(const Point&) const override;
-    int64_t DistPoint(const Point&) const;
+    float DistPoint(const Point&) const;
     bool CrossesSegment(const Segment&) const override;
     ~Segment() override;
   };
